backup_v1/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
groq==0.33.0
python-dotenv==1.0.0
werkzeug==3.0.1
python-dateutil==2.8.2
gunicorn==21.2.0
firebase-admin==6.5.0
redis==5.0.8

